Switch to build off of head of branch; revert tag to sha+arch; reorder events: clone, clear, drop in tests, commit, push/pull
