+++ .travis.yml
@@ -5,8 +5,7 @@ notifications:
 env:
   global:
     - RT_URL=github.com/big-ladder/kiva-test-results.git
-    - TEST_DIR=build/test/results
-    - RT_DIR=scripts/regress
+    - RT_DIR=build/test/results
     - secure: "Iu96o6rugFB7LdIGAY8+Yr2ChIZiLGmyWoX6CtXBfu72ZZZcI4G8QMLyh8ymPMBIoq6e7FQOiHyn0TGuPsutwOasnDIKvasP1Duy4hDmO8394uWnv0P2icWNBny7A+cv8mIby/BJ2hB0SlqkCGPi5+L+66FOTZMSw7KWoOxFIsd8RRha4nBY27flGrbcd+nTjyXXSlY6TLP/cpktucB600JCOwbHhV0RqYqd2IIAYKMJTWRVACu42oQqmxqO+T9HP0n0MlV2bbY7PgGEii79wieeaeaVyWdEWPN8fxEqeu8iXfYI3vijXo7IuDLm5kOpPoMFM+OyOWHzFJ0zsFkbJixE7V3F8qEVCiYFI+myGTrCnRSqD3jD0+GQjklj9P1+vmO0D9TBrH66z7YLL9t5E2lGOLJWnNI1uJoUKPTO+BrBvR4sOWuie3fBXCfMDbUr0N4uBfjnVSSAdHvjLf25pN3gkXEpptjSFjJ74blMVswHw9QlCQ4eXN+FGgoLhY2QpzPLpMFmH5ZX5R7r4jktqqNHy4rziM5R9Jxbdf9Gt4kjm0Uc4tEC4iohbTgGlDKIjZPcA2dP4QMCiuidXD80jl+KDSqaOHONB9urf2U0StFEmuJApbQ96gVo+a7/uEgb0izCnOv2V1BbGcQcgWrx7JkCipbae7Lsz/8Of4hqgXI="
 matrix:
   include:
@@ -35,6 +34,7 @@ script:
   - mkdir build
   - cd build
   - cmake -DPROCESSOR_ARCHITECTURE=$PROCESSOR_ARCHITECTURE ..
+  - ruby ../scripts/clone-and-mimic.rb
   - make -j
 
 after_success:
