+++ scripts/run-kiva.rb
@@ -19,6 +19,7 @@ def run_case(exe_path, in_file, weather_file, output_path)
   output_dir = File.dirname(output_path)
   puts("  ... output directory=#{output_dir}")
   puts("  ... exe path        =#{exe_path}")
+  FileUtils.rm_rf(output_dir)
   FileUtils.mkdir_p(output_dir) unless File.exist?(output_dir)
   FileUtils.cp(in_file, output_dir)
   FileUtils.cp(weather_file, output_dir)
